글 삭제시 권한체크 코드 추가, 테스트 케이스 수정

+++ src/main/java/com/restapi/domain/post/post/controller/ApiV1PostController.java
@@ -53,8 +53,21 @@ public class ApiV1PostController {
     @Transactional
     @DeleteMapping("/{id}")
     @Operation(summary = "삭제")
-    public RsData<PostDto> delete(@PathVariable long id) {
+    public RsData<PostDto> delete(
+            @PathVariable long id,
+            @NotBlank @Size(min = 2, max = 50) @RequestHeader("Authorization") String authorization
+    ) {
+        String apiKey = authorization.replace("Bearer ", "");
+
+        Member author = memberService.findByApiKey(apiKey)
+                .orElseThrow(() -> new ServiceException("401-1", "존재하지 않는 회원입니다."));
+
         Post post = postService.findById(id);
+
+        if (!author.equals(post.getAuthor())) {
+            throw new ServiceException("403-1", "글 삭제 권한이 없습니다.");
+        }
+
         postService.delete(post);
         return new RsData<>("200-1", "%d번 게시글이 삭제되었습니다.".formatted(id), new PostDto(post));
     }
